Arrage 3 numbers into ascending order
Write a function called arrangeThree that swaps the values of three
integers, given as three separate parameters.
This function should not have any return value, but, like swap() in the
lecture notes, moves the values
from one place to another. When the algorithm returns, the smallest
value should be in the first parameter,
and the largest is in the last (according to the order of your parameter
list).
For example, suppose you have three references whose values are in the
following order: 3,2,5. When your
function finishes, the references will have values in increasing order:
2,3,5.